+++ core/export/export.model.d.ts
@@ -1,6 +1,12 @@
 import {Resource} from '../resource/resource';
 
+/**
+ * A class to handle export settings.
+ *
+ */
+
 export declare class ExportModel {
 	constructor();
+
 	resource: Resource;
 }
