+++ core/export/json/json/json.d.ts
@@ -1,4 +1,4 @@
-import {ColumnModel} from '../column-type/column.model';
+import {ColumnModel} from '../../../column-type/column.model';
 
 export interface IOptions {
   structure: string;
