+++ core/export/xml/xml/xml.d.ts
@@ -1,4 +1,4 @@
-import {ColumnModel} from '../column-type/column.model';
+import {ColumnModel} from '../../../column-type/column.model';
 
 export declare class Xml {
 	write(rows: any[], columns?: ColumnModel[]): string;
