+++ core/import/import.model.d.ts
@@ -1,6 +1,12 @@
 import {Resource} from '../resource/resource';
 
+/**
+ * A class to handle import settings.
+ *
+ */
+
 export declare class ImportModel {
 	constructor();
+
 	resource: Resource;
 }
